Player and ship health pushes; timer cooldown maybe fixed

+++ Assets/Scripts/SpaceshipScripts/ShipHealth.cs
@@ -61,7 +61,7 @@ public class ShipHealth : Photon.MonoBehaviour {
 		if (stream.isWriting) {
 			stream.SendNext (health);
 		} else if (stream.isReading) {
-			health = (int)stream.ReceiveNext ();
+			health = (float)stream.ReceiveNext ();
 			tookDmg = true;
 		}
 	}
